requirements.txt
fastapi
pydantic
python-dotenv
psycopg2-binary
bcrypt
PyJWT
sqlalchemy
uvicorn
httpx
pwdlib
pwdlib[argon2]